Start of Multiplayer Changes
I added some new variables IsMultiplayer and Active player turn to help mange future multiplayer.
I remade some of the code for scoring to display player 1 and 2 instead of enemy if in multiplayer.
I made a GameUtilites static class fo future use that currently holds a method for comparing if a turn & team match.
Finally I updated Draggable script to use this new check instead of only early exiting if enemy as this does not work in multiplayer.